final tweaks and updated readme

+++ AStar.Tests/Helper.cs
@@ -123,7 +123,7 @@ namespace AStar.Tests
                 {
                     if (splitLevel[row][column] != closedCharacter)
                     {
-                        world[row, column] = int.Parse(splitLevel[row][column].ToString());
+                        world[row, column] = short.Parse(splitLevel[row][column].ToString());
                     }
                 }
             }
